+++ .changes/re-export-log.md
@@ -1,0 +1,6 @@
+---
+"log": minor
+"log-js": minor
+---
+
+Re-export the log crate.
